refactor(pi): extract the dev opener to dev.ts; create.ts = reusable ladder
Direction B: make the three CLI commands symmetric. createPiAgentFromWorkspace
(dev's opener) moves from create.ts into its own dev.ts, mirroring build.ts and
start.ts. create.ts now holds only the reusable assembly ladder (L0–L2 via
invoke.ts + L1/L2) and the engine assets (tools/prompt).

Rationale: createPiAgentFromWorkspace (dev) and createPiAgentFromArtifact (start)
are conceptual twins — open a source → resolve model/tools → inject K-wiring →
call L2 — but were placed asymmetrically (dev as an "L3 ladder rung" in create.ts,
start as a "sibling" in start.ts). They are command-posture openers, not pure
assembly primitives (they encode config/manifest precedence, sessions location,
--global-skills). Separating "reusable assembly (L0–L2)" from "command posture
(dev/build/start)" makes the structure legible: create.ts = ladder, dev/build/
start.ts = the three commands.

Pure refactor: public exports are unchanged (index.ts re-exports
createPiAgentFromWorkspace from dev.ts); no behavior change. Docs updated
(create.ts/dev.ts/start.ts headers, core-design §4/§7, fastagent.md, AGENTS.md map).
